requirements.txt
pycryptodomex
urllib3
pybase64
